refactor: merge kill-zombie-next.js into dev.js
Consolidate zombie cleanup logic directly into the dev script instead of
calling it as a separate process.
